make app one page using react-router-dom, update window animations and pathing
