+++ src/components/Window/window.module.scss
@@ -34,4 +34,54 @@
   top: 0;
   width: 100%;
   height: 100%;
+  background-color: $black-secondary;
+}
+
+.header {
+  position: sticky;
+  width: 100%;
+  left: 0;
+  z-index: 999;
+}
+
+.invisibleHeader {
+  visibility: hidden;
+  position: sticky;
+  width: 100%;
+  left: 0;
+  height: 52px;
+}
+
+.slideUpAnimation {
+  top: -60px;
+  animation-name: slideUp;
+	animation-iteration-count: 1;
+	animation-timing-function: cubic-bezier(0.25, 0.1, 0.5, 1);
+	animation-duration: 200ms;
+}
+
+.slideDownAnimation {
+  top: 0px;
+  animation-name: slideDown;
+	animation-iteration-count: 1;
+	animation-timing-function: cubic-bezier(0.25, 0.1, 0.5, 1);
+	animation-duration: 200ms;
+}
+
+@keyframes slideDown {
+	0% {
+    transform: translateY(-60px);
+	}
+	100% {
+    top: 0;
+	}
+}
+
+@keyframes slideUp {
+	0% {
+    top: 0;
+	}
+	100% {
+    transform: translateY(-60px);
+	}
 }
