+++ src/package.json
@@ -15,6 +15,7 @@
     "plaiceholder": "^2.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-spring": "^9.2.4",
     "react-transition-group": "^4.4.2",
     "sass": "^1.36.0",
