+++ src/pages/_app.tsx
@@ -1,21 +1,13 @@
 import { AppProps } from 'next/app'
-import { Router } from 'next/router';
-import { useEffect } from 'react'
 
 import '../styles/global.scss'
 
 function App({ Component, pageProps }: AppProps) {
-  useEffect(() => {
-    Router.events.on('routeChangeComplete', () => {
-      window.scroll({
-        top: 1,
-        left: 0,
-        behavior: 'smooth'
-      })
-    })
-  }, [])
-
-  return <Component {...pageProps} />
+  return (
+    <div suppressHydrationWarning>
+      {typeof window === 'undefined' ? null : <Component {...pageProps} />}
+    </div>
+  )
 }
 
 export default App
