+++ src/views/homepage/homepage.module.scss
@@ -1,6 +1,6 @@
-@import '../styles/breakpoints.module.scss';
-@import '../styles/colors.module.scss';
-@import '../styles/fonts.module.scss';
+@import '../../styles/breakpoints.module.scss';
+@import '../../styles/colors.module.scss';
+@import '../../styles/fonts.module.scss';
 
 .sideBar {
   display: flex;
@@ -159,5 +159,28 @@
 	}
 	100% {
 		opacity: 1;
+	}
+}
+
+.issueTransition {
+  background-color: $black-secondary;
+  width: 100%;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  height: 100%;
+
+  animation-name: slideUp;
+	animation-iteration-count: 1;
+	animation-timing-function: cubic-bezier(0.25, 0.1, 0.5, 1);
+	animation-duration: 400ms;
+}
+
+@keyframes slideUp {
+	0% {
+		height: 0;
+	}
+	100% {
+    height: 100%;
 	}
 }
