+++ src/views/issue/issue.tsx
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { CSSTransition } from 'react-transition-group'
-import { useRouter } from 'next/router'
+import { useNavigate, useLocation } from "react-router-dom";
 import { useEffect, useState } from 'react'
 
 import Window from '../../components/Window/window'
@@ -11,8 +11,6 @@ import styles from './issue.module.scss'
 import Issue_0 from '../../copy/issue/0'
 import useBrowserSize from '../../common/hooks/useWindowSize'
 import { animated, useSprings } from 'react-spring'
-
-const ISSUES = [Issue_0]
 
 type Overview = {
   concept: string,
@@ -25,22 +23,25 @@ export default function Issue({ name, overview }: {
   name: string,
   overview: Overview,
 }) {
-  const [issueState, setIssueState] = useState<IssueView>({ state: "info" })
-  const [isOpen, setIsOpen] = useState(true)
-  const router = useRouter()
+  const navigate = useNavigate()
+  const location = useLocation()
   const windowDimension = useBrowserSize()
+
+  const initialState = location.state?.enter === 'rotate' ? 'rotate' : 'info'
+  const [isOpen, setIsOpen] = useState(true)
+  const [issueState, setIssueState] = useState<IssueView>({ state: initialState })
   const tooNarrow = windowDimension && windowDimension.width < 540
 
   const handleMinimize = () => {
     setIsOpen(false)
-    setTimeout(() => router.back(), 200)
+    navigate('/')
   }
   const handleEnter = () => {
     setIssueState(tooNarrow ? { state: "rotate" } : { state: "project" })
   }
   const handleTransitionDone = () => {
     if ("transitionTo" in issueState) {
-      router.push(Issue_0_entries[issueState.transitionTo][1].path)
+      navigate(Issue_0_entries[issueState.transitionTo][1].path, { replact: true })
     }
   }
 
@@ -83,6 +84,7 @@ export default function Issue({ name, overview }: {
             isOpen={isOpen}
             isFullscreen
             hasContentPadding={false}
+            animationDuration={50}
           >
             <div className={styles.container}>
               <div className={styles.content}>
@@ -118,18 +120,6 @@ export default function Issue({ name, overview }: {
   )
 }
 
-export async function getStaticPaths() {
-  const paths = ISSUES.map((issue, index) => ({
-    params: { id: index.toString() },
-  }))
-  return { paths, fallback: false }
-}
-
-export async function getStaticProps({ params }) {
-  const issue = ISSUES[params.id]
-  return { props: { name: issue.name, overview: issue.overview } }
-}
-
 const RadialStripes = () => (
   <path id="Vector_3" d="M736.222 -329.524C745.863 -329.524 755.323 -329.342 764.601 -329.16L736.222 645.631C735.676 645.631 736.767 645.631 736.222 645.631L707.297 -329.16C716.938 -329.342 726.58 -329.524 736.222 -329.524ZM821.724 -325.877C840.643 -324.236 859.927 -321.865 878.664 -319.129L736.04 645.631C735.494 645.631 736.586 645.631 736.04 645.631L821.724 -325.877ZM650.174 -325.695L736.222 645.631C735.858 645.631 736.586 645.631 736.222 645.631L592.869 -318.947C611.971 -321.682 631.072 -324.053 650.174 -325.695ZM935.787 -308.916C954.525 -304.904 973.444 -300.527 992 -295.42L736.222 645.631C735.858 645.449 736.586 645.631 736.222 645.631L935.787 -308.916ZM536.111 -308.734L736.222 645.631C735.858 645.631 736.586 645.449 736.222 645.631L479.898 -295.056C498.453 -300.162 517.373 -304.722 536.111 -308.734ZM1047.3 -278.46C1065.5 -272.259 1083.69 -265.511 1101.52 -258.216L736.222 645.631C735.858 645.449 736.586 645.813 736.222 645.631L1047.3 -278.46ZM424.594 -278.277L736.222 645.449C735.858 645.631 736.586 645.266 736.222 645.449L370.382 -258.034C388.21 -265.146 406.402 -272.077 424.594 -278.277ZM1154.64 -234.872C1171.92 -226.665 1189.2 -217.729 1205.94 -208.428L736.222 645.631C735.858 645.449 736.586 645.813 736.222 645.631L1154.64 -234.872ZM317.626 -234.69L736.222 645.449C735.858 645.631 736.586 645.266 736.222 645.449L266.325 -208.246C283.061 -217.547 300.343 -226.301 317.626 -234.69ZM1255.78 -179.066C1271.97 -168.853 1287.98 -157.91 1303.63 -146.785L736.222 645.631C735.858 645.449 736.586 645.813 736.222 645.631L1255.78 -179.066ZM216.661 -178.701L736.222 645.631C735.858 645.813 736.586 645.266 736.222 645.631L169.18 -146.603C184.643 -157.728 200.652 -168.488 216.661 -178.701ZM1349.29 -111.77C1364.21 -99.7328 1378.76 -86.9666 1392.95 -74.018L736.222 645.631C735.858 645.266 736.586 645.996 736.222 645.631L1349.29 -111.77ZM123.518 -111.587L736.222 645.631C735.858 645.813 736.586 645.266 736.222 645.631L79.8574 -73.8356C93.8652 -86.7842 108.601 -99.5504 123.518 -111.587ZM1434.06 -33.8957C1447.34 -20.2176 1460.44 -5.81006 1472.99 8.59753L736.222 645.631C735.858 645.266 736.586 645.996 736.222 645.631L1434.06 -33.8957ZM38.5617 -33.7133L736.222 645.631C735.858 645.996 736.586 645.266 736.222 645.631L-0.187027 8.77991C12.1835 -5.81003 25.2816 -20.0352 38.5617 -33.7133ZM1509.01 53.2792C1520.66 68.5987 1531.94 84.2829 1542.67 100.149L736.222 645.631C736.04 645.266 736.586 645.996 736.222 645.631L1509.01 53.2792ZM-36.5709 53.4616L736.04 645.449C735.676 645.813 736.404 645.084 736.04 645.449L-70.2259 100.149C-59.3107 84.4652 -48.0318 68.7811 -36.5709 53.4616ZM1573.41 148.843C1583.24 165.257 1592.52 182.4 1601.25 199.361L736.222 645.631C736.04 645.266 736.404 645.996 736.222 645.631L1573.41 148.843ZM-100.97 149.208L736.222 645.631C736.04 645.996 736.404 645.266 736.222 645.631L-128.804 199.908C-120.072 182.765 -110.612 165.804 -100.97 149.208ZM1626.17 251.52C1633.99 269.028 1641.27 287.083 1648 305.138L736.222 645.631C736.04 645.266 736.404 645.996 736.222 645.631L1626.17 251.52ZM-153.727 252.25L736.222 645.631C736.04 645.996 736.404 645.266 736.222 645.631L-175.557 306.05C-168.826 287.995 -161.549 269.758 -153.727 252.25ZM1666.56 360.033C1672.2 378.453 1677.29 397.237 1681.84 416.022L736.222 645.631C736.222 645.266 736.404 645.996 736.222 645.631L1666.56 360.033ZM-194.113 361.127L736.222 645.631C736.04 645.996 736.404 645.266 736.222 645.631L-209.212 416.934C-204.846 398.331 -199.752 379.547 -194.113 361.127ZM1693.66 472.74C1697.12 491.707 1699.85 510.856 1702.21 530.006L736.222 645.631C736.222 645.266 736.222 645.996 736.222 645.631L1693.66 472.74ZM-221.037 473.652L736.222 645.631C736.222 645.996 736.222 645.266 736.222 645.631L-229.587 530.917C-227.404 511.768 -224.493 492.619 -221.037 473.652ZM1707.31 587.271C1708.4 606.238 1708.94 625.57 1708.94 644.537L736.04 645.631C736.04 645.084 736.04 646.178 736.04 645.631L1707.31 587.271ZM-234.681 588.183L736.04 645.631C736.04 646.178 736.04 645.084 736.04 645.631L-236.5 646.725V645.631C-236.318 626.482 -235.772 607.15 -234.681 588.183ZM736.222 645.631L1707.49 701.802C1706.4 720.952 1704.76 740.283 1702.58 759.25L736.222 645.631C736.222 645.084 736.222 645.996 736.222 645.631ZM736.222 645.631C736.222 646.178 736.222 645.084 736.222 645.631L-229.223 762.351C-231.406 743.384 -233.225 724.052 -234.317 704.903L736.222 645.631ZM736.222 645.631L1694.21 816.516C1690.93 835.483 1686.93 854.632 1682.38 873.234L736.222 645.631C736.222 645.266 736.222 645.996 736.222 645.631ZM736.222 645.631C736.222 645.996 736.04 645.266 736.222 645.631L-208.848 876.152C-213.396 857.55 -217.399 838.401 -220.855 819.434L736.222 645.631ZM736.222 645.631L1667.28 929.223C1661.64 947.643 1655.46 966.245 1648.73 984.3L736.222 645.631C736.404 645.266 736.04 645.996 736.222 645.631ZM736.222 645.631C736.222 645.996 736.04 645.266 736.222 645.631L-174.829 987.036C-181.56 968.981 -187.746 950.561 -193.385 932.141L736.222 645.631ZM736.222 645.631L1626.9 1038.1C1619.26 1055.79 1610.89 1073.3 1601.97 1090.44L736.222 645.631C736.404 645.266 736.04 645.996 736.222 645.631ZM736.222 645.631C736.404 645.996 736.04 645.266 736.222 645.631L-128.076 1092.81C-136.808 1075.85 -145.177 1058.16 -152.999 1040.65L736.222 645.631ZM736.222 645.631L1574.32 1141.14C1564.68 1157.74 1554.31 1174.15 1543.58 1190.02L736.222 645.631C736.404 645.266 736.04 645.996 736.222 645.631ZM736.222 645.631C736.404 645.996 735.858 645.266 736.222 645.631L-69.3163 1192.02C-80.0495 1176.16 -90.4189 1159.74 -100.061 1143.33L736.222 645.631ZM736.222 645.631L1509.92 1236.89C1498.28 1252.03 1486.27 1267.16 1473.72 1281.57L736.222 645.631C736.586 645.266 735.858 645.996 736.222 645.631ZM736.222 645.631C736.586 645.996 735.858 645.266 736.222 645.631L0.722565 1283.39C-11.6479 1268.99 -24.0184 1254.03 -35.4793 1238.71L736.222 645.631ZM736.222 645.631L1434.97 1324.25C1421.69 1337.92 1407.87 1351.42 1393.86 1364.37L736.222 645.631C736.586 645.266 735.858 645.996 736.222 645.631ZM736.222 645.631C736.586 645.996 735.858 645.266 736.222 645.631L80.767 1366.01C66.7592 1353.06 52.7514 1339.57 39.4713 1325.89L736.222 645.631ZM736.222 645.631L1350.2 1402.12C1335.46 1414.16 1320 1426.01 1304.54 1437.14L736.222 645.631C736.586 645.266 735.858 645.813 736.222 645.631ZM736.222 645.631C736.586 645.813 735.858 645.266 736.222 645.631L170.089 1438.41C154.626 1427.29 139.163 1415.43 124.428 1403.4L736.222 645.631ZM736.222 645.631L1256.87 1469.42C1240.68 1479.63 1223.95 1489.66 1207.21 1498.96L736.222 645.631C736.586 645.449 735.858 645.813 736.222 645.631ZM736.222 645.631C736.586 645.813 735.858 645.449 736.222 645.631L267.78 1500.24C251.043 1490.94 234.307 1481.09 218.116 1470.87L736.222 645.631ZM736.222 645.631C736.586 645.813 735.858 645.449 736.222 645.631L372.02 1549.84C354.192 1542.73 336.545 1534.89 319.081 1526.5L736.222 645.631ZM736.222 645.631L1156.09 1525.4C1138.81 1533.61 1120.98 1541.64 1103.33 1548.75L736.222 645.631C736.586 645.449 735.858 645.813 736.222 645.631ZM736.222 645.631C736.586 645.631 735.858 645.449 736.222 645.631L481.899 1586.86C463.343 1581.76 444.787 1576.1 426.595 1570.09L736.222 645.631ZM736.222 645.631L1049.12 1568.99C1030.93 1575.19 1012.38 1580.85 993.819 1585.95L736.222 645.631C736.586 645.449 735.858 645.631 736.222 645.631ZM736.222 645.631L937.788 1599.63C919.051 1603.64 899.949 1607.11 880.847 1609.84L736.222 645.631C736.586 645.449 735.858 645.631 736.222 645.631ZM736.222 645.631C736.586 645.631 735.858 645.449 736.222 645.631L594.87 1610.39C575.951 1607.66 556.667 1604.19 537.93 1600.36L736.222 645.631ZM736.222 645.631L823.725 1616.77C804.805 1618.42 785.522 1619.69 766.42 1620.24L736.222 645.631C736.767 645.631 735.676 645.631 736.222 645.631ZM736.222 645.631C736.767 645.631 735.676 645.631 736.222 645.631L709.298 1620.24C690.196 1619.69 670.913 1618.6 651.993 1616.96L736.222 645.631Z" fill="#21763B" />
 )
@@ -149,8 +139,6 @@ const OverviewGraphics = ({ width, height, setTransitionTo, transitionTo }) => {
   const [menuOpen, setMenuOpen] = useState(false)
 
   const [hoverId, setHoverId] = useState(-1)
-
-  const router = useRouter()
 
   const w = Math.max(0, unit(1160, 1472, width))
   const h = Math.max(0, unit(700, 1080, height))
